abstract_shadow.hpp etc.
