+++ rtlib/signature.hpp
